feature: Adding compatibility to IPv6
Using socket.create_connection instead of socket.connect
This allow the function to work for both IPv4 and IPv6 addresses vs
the current current implementation that support only IPv4 (AF_INET).
also if host is a non-numeric hostname, it will try to resolve it for both
AF_INET and AF_INET6,
and then try to connect to all possible addresses in turn until a
connection succeeds.