feature: Adding a new function get_free_port() to nettools
because picking a random port is not a good idea - this function simplify
the process of getting available port by letting OS pick free one for you.
tought this function can be useful in other parts
of the project it added in this commit to nettools because its code was
needed repeatedly in the unittest of nettools.